refactor: move helper classes to internal package and update visibility of methods
