+++ src/main/java/com/github/eventmanager/internal/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains helper classes that are used to perform various operations in the application.
  */
-package com.github.eventmanager.helpers;
+package com.github.eventmanager.internal;
