+++ src/main/java/module-info.java
@@ -12,6 +12,6 @@ module com.github.eventmanager {
     exports com.github.eventmanager.filehandlers.config;
     exports com.github.eventmanager.formatters;
     exports com.github.eventmanager.processors;
-    exports com.github.eventmanager.helpers;
+    exports com.github.eventmanager.internal;
     exports com.github.eventmanager.outputs;
 }
